backend/pyproject.toml
[tool.poetry]
name = "squatofadt"
version = "0.1.0"
description = "Squatten of adten"
authors = ["Lars van Rhijn <larsvanrhijn@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.3"
uWSGI = "^2.0.19"
psycopg2-binary = "^2.8.6"
djangorestframework = "^3.12.4"
django-oauth-toolkit = "^1.5.0"
PyYAML = "^5.4.1"
uritemplate = "^3.0.1"
django-cors-headers = "^3.7.0"

[tool.poetry.dev-dependencies]
pydocstyle = "^5.0.2"
black = "^19.10b0"
flake8 = "^3.7.9"
coverage = "^5.5"

[tool.black]
line-length = 119
target-version = ["py37"]
exclude = '''
/(
  migrations
  | website/squatofadt/settings
)/
'''

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
